\name{class}

\alias{class}

\docType{data}

\title{
Class labels of samples
}

\description{
Class labels of samples
}

\usage{data(class)}

\format{
  The format is:
 num [1, 1:4000] 2 1 1 2 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4000] "Sample1" "Sample2" "Sample3" "Sample4" ...
}

\references{
None
}

\examples{
data(class)
}

\keyword{datasets}
