\name{Edges}

\alias{Edges}

\docType{data}

\title{
Edges of constructed graphs
}

\description{
Edges of constructed graphs
}

\usage{data(Edges)}

\format{
  The format is:
 chr [1:45, 1:2] "rs00008:rs00019" "rs00005:rs00017" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "From" "To"
}

\references{
None
}

\examples{
data(Edges)
}

\keyword{datasets}
