\name{FiveEffect}

\alias{FiveEffect}

\docType{data}

\title{
Interaction Effects of five-SNP combinations
}

\description{
Interaction Effects of five-SNP combinations
}

\usage{data(FiveEffect)}

\format{
  The format is:
 num [1, 1] 0
}

\references{
None
}

\examples{
data(FiveEffect)
}

\keyword{datasets}
