\name{FourEffect}

\alias{FourEffect}

\docType{data}

\title{
Interaction Effects of four-SNP combinations
}

\description{
Interaction Effects of four-SNP combinations
}

\usage{data(FourEffect)}

\format{
  The format is:
 num [1, 1] 0
}

\references{
None
}

\examples{
data(FourEffect)
}

\keyword{datasets}
