\name{InstallPackage}

\alias{InstallPackage}

\title{
Install CINOEDV used packages
}

\description{
Install CINOEDV used packages, including R.matlab, igraph, ggplot2, and reshape2.
}

\usage{
InstallPackage()
}

\value{
None
}

\references{
None
}

\author{
Junliang Shang \email{shangjunliang110@163.com}
}

\examples{
InstallPackage()
}

