\name{pts}

\alias{pts}

\docType{data}

\title{
SNP data
}

\description{
SNP data
}

\usage{data(pts)}

\format{
  The format is:
 int [1:4000, 1:20] 1 2 1 3 1 2 2 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:4000] "Sample1" "Sample2" "Sample3" "Sample4" ...
  ..$ : chr [1:20] "SNP1" "SNP2" "SNP3" "SNP4" ...
}

\references{
None
}

\examples{
data(pts)
}

\keyword{datasets}
