\name{SingleEffect}

\alias{SingleEffect}

\docType{data}

\title{
Main Effects of SNPs
}

\description{
Main Effects of SNPs
}

\usage{data(SingleEffect)}

\format{
  The format is:
 chr [1:20, 1:2] "rs00013" "rs00010" "rs00006" "rs00001" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "SNP" "Value"
}

\references{
None
}

\examples{
data(SingleEffect)
}

\keyword{datasets}
