\name{TestPSOParameters}

\alias{TestPSOParameters}

\title{
Check the PSO parameters
}

\description{
This function Check the PSO parameters 'Population' and 'Iteration'.
}

\usage{
TestPSOParameters(Population, Iteration)
}

\arguments{
  \item{Population}{
numeric. The number of particles.
}
  \item{Iteration}{
numeric. The number of iterations.
}
}

\value{
None
}

\references{
None
}

\author{
Junliang Shang \email{shangjunliang110@163.com}
}

\examples{
Population <- 1000
Iteration <- 50
TestPSOParameters(Population, Iteration)
}
