\name{ThreeEffect}

\alias{ThreeEffect}

\docType{data}

\title{
Interaction Effects of three-SNP combinations
}

\description{
Interaction Effects of three-SNP combinations
}

\usage{data(ThreeEffect)}

\format{
  The format is:
 chr [1:1137, 1:4] "rs00002" "rs00008" "rs00004" "rs00001" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "SNP1" "SNP2" "SNP3" "Value"
}

\references{
None
}

\examples{
data(ThreeEffect)
}

\keyword{datasets}
