\name{TwoEffect}

\alias{TwoEffect}

\docType{data}

\title{
Interaction Effects of two-SNP combinations
}

\description{
interaction Effects of two-SNP combinations
}

\usage{data(TwoEffect)}

\format{
  The format is:
 chr [1:187, 1:3] "rs00008" "rs00005" "rs00013" "rs00005" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3] "SNP1" "SNP2" "Value"
}

\references{
None
}

\examples{
data(TwoEffect)
}

\keyword{datasets}
