\name{Vertices}

\alias{Vertices}

\docType{data}

\title{
Vertices of constructed graphs
}

\description{
Vertices of constructed graphs, including real vertices and virtual vertices.
}

\usage{data(Vertices)}

\format{
  The format is:
 chr [1:37, 1:3] "rs00002" "rs00008" "rs00004" "rs00001" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3] "id" "value" "label"
}

\references{
None
}

\examples{
data(Vertices)
}

\keyword{datasets}
